requirements.txt
torch==1.12.1
transformers==4.20.1
Pillow==9.1.0
matplotlib
